+++ packages/google_mobile_ads/android/src/main/java/io/flutter/plugins/googlemobileads/FlutterAdSize.java
@@ -32,6 +32,10 @@ class FlutterAdSize {
     AdSize getLandscapeAnchoredAdaptiveBannerAdSize(Context context, int width) {
       return AdSize.getLandscapeAnchoredAdaptiveBannerAdSize(context, width);
     }
+
+    AdSize getCurrentOrientationAnchoredAdaptiveBannerAdSize(Context context, int width) {
+      return AdSize.getCurrentOrientationAnchoredAdaptiveBannerAdSize(context, width);
+    }
   }
 
   static class AnchoredAdaptiveBannerAdSize extends FlutterAdSize {
@@ -39,27 +43,23 @@ class FlutterAdSize {
 
     @NonNull
     private static AdSize getAdSize(
-        @NonNull Context context,
-        @NonNull AdSizeFactory factory,
-        @NonNull String orientation,
-        int width) {
+        @NonNull Context context, @NonNull AdSizeFactory factory, String orientation, int width) {
       final AdSize adSize;
-      if (orientation.equals("portrait")) {
+      if (orientation == null) {
+        adSize = factory.getCurrentOrientationAnchoredAdaptiveBannerAdSize(context, width);
+      } else if (orientation.equals("portrait")) {
         adSize = factory.getPortraitAnchoredAdaptiveBannerAdSize(context, width);
       } else if (orientation.equals("landscape")) {
         adSize = factory.getLandscapeAnchoredAdaptiveBannerAdSize(context, width);
       } else {
         throw new IllegalArgumentException(
-            "Orientation should be 'portrait' or 'landscape': " + orientation);
+            "Orientation (optional) should be 'portrait' or 'landscape': " + orientation);
       }
       return adSize;
     }
 
     AnchoredAdaptiveBannerAdSize(
-        @NonNull Context context,
-        @NonNull AdSizeFactory factory,
-        @NonNull String orientation,
-        int width) {
+        @NonNull Context context, @NonNull AdSizeFactory factory, String orientation, int width) {
       super(getAdSize(context, factory, orientation, width));
       this.orientation = orientation;
     }